Add new structure + refactor for integration tests #282
