+++ test/integration.old/__fast-price-test.ts__
