+++ test/integration/01-setup/01-metadata.spec.ts
@@ -1,0 +1,7 @@
+import { DbMetadata } from "src/schema"
+
+describe("DbMetadata", () => {
+  it("sets min version for graphql", async () => {
+    await DbMetadata.create({ minBuildNumber: 200, lastBuildNumber: 200 })
+  })
+})
